Adds Links, refactors Menu.
