+++ src/components/Link/Link.styles.ts
@@ -1,4 +1,4 @@
-import { createStyles, makeStyles, Theme } from '@material-ui/core';
+import { createStyles, makeStyles } from '@material-ui/core';
 
 export default makeStyles(() =>
   createStyles({
