--- src/pages/MyPosts/MyPosts.styles.ts
@@ -1,14 +1,0 @@
-import { createStyles, makeStyles, Theme } from '@material-ui/core';
-
-export default makeStyles((theme: Theme) =>
-  createStyles({
-    root: {
-      paddingTop: 24,
-    },
-    grid: {
-      [theme.breakpoints.down('sm')]: {
-        display: 'none',
-      },
-    },
-  })
-);
